+++ src/internal/MatcherPrinter.ts
@@ -1,7 +1,7 @@
 import { Matcher, MatchingLogic } from "../Matchers";
 import { Concat, isNamedMatcher, MatchStep } from "../matchers/Concat";
-import { isBreak } from "../matchers/snobol/Break";
 import { isLiteral } from "../Primitives";
+import { isBreak } from "./Break";
 import Match = Chai.Match;
 
 /**