+++ test/integration/Hcl.ts
@@ -4,9 +4,9 @@ import { Term } from "../../src/Matchers";
 import { Microgrammar } from "../../src/Microgrammar";
 import { Integer, LowercaseBoolean } from "../../src/Primitives";
 
-import { Break } from "../../src/matchers/snobol/Break";
-
 import { Alt } from "../../src/Ops";
+
+import { takeUntil } from "../../src/matchers/skip/Skip";
 
 const sample = `# An AMI
 variable "ami" {
@@ -31,7 +31,7 @@ describe("Parsing HCL", () => {
     // Function so as not to trigger eager loading, which can fail tests
     function hclString() {
         return Microgrammar.fromString(`"\${stuffUntilNextQuote}"`, {
-            stuffUntilNextQuote: new Break('"'),
+            stuffUntilNextQuote: takeUntil('"'),
         });
     }